+++ src/main/java/de/maxhenkel/camera/entities/ImageEntityRenderState.java
@@ -2,9 +2,9 @@ package de.maxhenkel.camera.entities;
 
 import net.minecraft.client.renderer.entity.state.EntityRenderState;
 import net.minecraft.core.Direction;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.phys.AABB;
 
-import javax.annotation.Nullable;
 import java.util.UUID;
 
 public class ImageEntityRenderState extends EntityRenderState {
@@ -13,9 +13,11 @@ public class ImageEntityRenderState extends EntityRenderState {
     public int frameWidth;
     public int frameHeight;
     public Direction facing;
-    @Nullable
-    public UUID imageUUID;
+    public ImageState imageState;
     public UUID imageEntityUUID;
     public AABB imageBoundingBox;
+
+    public record ImageState(UUID imageId, float imageRatio, ResourceLocation resourceLocation) {
+    }
 
 }